requirements.txt
ghostscript==0.7
python-dotenv==1.0.0
langchain==0.0.184
opencv-python==4.10.0.84
opencv-python-headless==4.10.0.84
camelot-py==0.11.0 
matplotlib==3.10.0
pytesseract==0.3.13
arabic-reshaper==3.0.0
python-bidi==0.6.3
joblib==1.4.2
pdf2image==1.17.0
ghostscript==0.7
watchdog==3.0.0
pymongo==4.6.1
fastapi==0.109.0
uvicorn==0.27.0
python-multipart==0.0.6
jinja2==3.1.3